cpu_input_boost: Rework display state awareness logic for OnePlus 7 devices
Currently, cpu_input_boost fails to engage when the device is
initially powered on until the display is manually cycled off/on.
Additionally, it occasionally fails to re-engage following
subsequent display power cycles. In order to ensure proper
functionality, we need to be more explicit about which display
states the driver is checking for.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: Zlatan Radovanovic <zlatan.radovanovic@fet.ba>
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>